feat: implement createSupplyUseCase

+++ app/src/main/java/com/bulletapps/candypricer/data/parameters/CreateSupplyParameters.kt
@@ -6,9 +6,9 @@ data class CreateSupplyParameters(
     @SerializedName("name")
     var name: String,
     @SerializedName("quantity")
-    var quantity: Double,
+    var quantity: Int, // TODO should be Double
     @SerializedName("value")
     var price: Double,
     @SerializedName("unit_id")
-    var unitId: Double,
+    var unitId: Int,
 )
